minor fixes

+++ Lab03 - Rasterization/Lab03 - Rasterization/Shapes.cs
@@ -74,7 +74,7 @@ namespace Lab03___Rasterization
 
         protected double DistanceBetween(Point p1, Point p2)
         {
-            return (int)Math.Round(Math.Sqrt(Math.Pow((p2.X - p1.X), 2) + Math.Pow((p2.Y - p1.Y), 2)));
+            return Math.Round(Math.Sqrt(Math.Pow((p2.X - p1.X), 2) + Math.Pow((p2.Y - p1.Y), 2)));
         }
 
         protected double DistanceFromLine(Point onLine1, Point onLine2, Point exterior)
